build fixes
- boost rejects some float, simply casting it to int should work
- add some default constructor
- moving DLLEXPORT to cpp makes shipping generated config.h in
  include/sbe uneccessary. People on SO say this works in cpps too:
  https://stackoverflow.com/a/557365

+++ src/util/TickControl.cpp
@@ -50,7 +50,7 @@ namespace sbe
 			MsToNextTick = MaxTickDuration - LastTickDuration;
 			if (Lag <= MsToNextTick)
 			{
-				boost::this_thread::sleep(boost::posix_time::milliseconds(MsToNextTick - Lag));
+				boost::this_thread::sleep(boost::posix_time::milliseconds( (int) (MsToNextTick - Lag)));
 				Lag = 0;
 			}
 			else
